refactor: move Prismic clients under src/lib/prismic
